Update build script

+++ build.sh
@@ -1,22 +1,19 @@
 #!/bin/bash
 
 REPO_URL="https://gitlab.com/devkcud/pun.git"
-TEMP_FOLDER="/tmp/pun_src"
+TEMP_FOLDER="/tmp/pun/"
 BINARY_FOLDER="/usr/local/bin"
 
-## Temporary folder creation
+## Creating temporary folder
 echo "Creating '$TEMP_FOLDER'."
 
-if [ -d "$TEMP_FOLDER" ]; then
-  rm -rf $TEMP_FOLDER
-fi
-
+if [ -d "$TEMP_FOLDER" ]; then rm -rf $TEMP_FOLDER; fi
 mkdir $TEMP_FOLDER
 
 ## Cloning repo
 echo "Cloning repo from '$REPO_URL'"
-git clone $REPO_URL $TEMP_FOLDER/pun -q
-cd $TEMP_FOLDER/pun || exit 1
+git clone $REPO_URL $TEMP_FOLDER -q
+cd $TEMP_FOLDER || exit 1
 
 ## Building
 echo "Running cargo build."
@@ -24,12 +21,11 @@ cargo build --release --quiet
 
 ## Copying binary file
 echo "Copying to '$BINARY_FOLDER'"
+sudo cp ./target/release/pun $BINARY_FOLDER/pun
 
-if [ -f "$BINARY_FOLDER" ]; then
-  sudo rm $BINARY_FOLDER/pun
-fi
-
-sudo cp ./target/release/pun $BINARY_FOLDER
+# Removing temporary folder
+echo "Removing '$TEMP_FOLDER'"
+if [ -d "$TEMP_FOLDER" ]; then rm -rf $TEMP_FOLDER; fi
 
 echo "Done."
